requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
boto3==1.35.24
python-dotenv==1.0.1
pillow==10.4.0
python-multipart==0.0.12
numpy==1.26.4
scipy==1.14.1
opencv-python==4.10.0.84
mediapipe==0.10.14

